refactor(observability): flatten lib structure
