+++ libs/observability/src/index.ts
@@ -1,7 +1,7 @@
 export * from './lib/interceptors/logging.interceptor';
-export * from './lib/open-telemetry/tracing';
-export * from './lib/open-telemetry/utils/inject-trace-into-data';
-export * from './lib/open-telemetry/wrappers/with-job-tracing.wrapper';
-export * from './lib/winston/config/winston-config.factory';
-export * from './lib/winston/formatters/otel-trace-log.formatter';
-export * from './lib/winston/formatters/secret-masking.formatter';
+export * from './lib/configs/tracing';
+export * from './lib/utils/inject-trace-into-data';
+export * from './lib/wrappers/with-job-tracing.wrapper';
+export * from './lib/configs/winston-config.factory';
+export * from './lib/formatters/otel-trace-log.formatter';
+export * from './lib/formatters/secret-masking.formatter';
