+++ libs/observability/src/lib/configs/winston-config.factory.ts
@@ -1,4 +1,4 @@
-import { otelTraceLogFormatter } from 'libs/observability/src/lib/winston/formatters/otel-trace-log.formatter';
+import { otelTraceLogFormatter } from '../formatters/otel-trace-log.formatter';
 import { extractBoolEnv, extractStrEnvWithFallback } from 'shared-kernel';
 import * as winston from 'winston';
