+++ libs/observability/src/lib/formatters/otel-trace-log.formatter.ts
