+++ libs/observability/src/lib/formatters/secret-masking.formatter.ts
